+++ lib/index.d.ts
@@ -179,6 +179,8 @@ declare namespace Realm {
         readonly type: PropertyType;
         readonly optional: boolean;
 
+        description(): string;
+
         /**
          * @returns boolean
          */
@@ -209,6 +211,7 @@ declare namespace Realm {
          * @returns Results<T>
          */
         subscribe(subscriptionName?: string): Realm.Sync.Subscription;
+        subscribe(options?: Realm.Sync.SubscriptionOptions): Realm.Sync.Subscription;
 
         /**
          * @returns Results
@@ -318,6 +321,12 @@ declare namespace Realm.Sync {
     interface SerializedTokenUser {
         server: string;
         adminToken: string;
+    }
+
+    interface SubscriptionOptions {
+        name: string;
+        update?: boolean;
+        timeToLive: number;
     }
 
     class AdminCredentials extends Credentials {
@@ -712,8 +721,8 @@ interface ProgressPromise extends Promise<Realm> {
 }
 
 interface NamedSubscription {
-    name: string,
-    objectType: string,
+    readonly name: string,
+    readonly objectType: string,
     query: string
     readonly state: SubscriptionState;
     readonly error: string;