+++ src/js_results.hpp
@@ -317,14 +317,36 @@ void ResultsClass<T>::subscribe(ContextType ctx, ObjectType this_object, Argumen
     auto sync_config = realm->config().sync_config;
 
     util::Optional<std::string> subscription_name;
+    bool update = false;
+    util::Optional<int64_t> ttl = none;
     if (args.count == 1) {
-        subscription_name = util::Optional<std::string>(Value::validated_to_string(ctx, args[0]));
+        if (Value::is_string(ctx, args[0])) {
+            subscription_name = util::Optional<std::string>(Value::validated_to_string(ctx, args[0]));
+        } else {
+            ObjectType options_object = Value::validated_to_object(ctx, args[0]);
+            ValueType name_value = Object::get_property(ctx, options_object, "name");
+            if (Value::is_undefined(ctx, name_value)) {
+                throw std::logic_error("A 'name' must be set.");
+            } else {
+                subscription_name = util::Optional<std::string>(Value::validated_to_string(ctx, name_value));
+            }
+
+            ValueType update_value = Object::get_property(ctx, options_object, "update");
+            if (!Value::is_undefined(ctx, update_value)) {
+                update = Value::validated_to_boolean(ctx, update_value, "update");
+            }
+
+            ValueType ttl_value = Object::get_property(ctx, options_object, "timeToLive");
+            if (!Value::is_undefined(ctx, ttl_value)) {
+                ttl = util::Optional<int64_t>(Value::validated_to_number(ctx, ttl_value, "timeToLive"));
+            }
+        }
     }
     else {
         subscription_name = util::none;
     }
 
-    auto subscription = partial_sync::subscribe(*results, subscription_name);
+    auto subscription = partial_sync::subscribe(*results, subscription_name, ttl, update);
     return_value.set(SubscriptionClass<T>::create_instance(ctx, std::move(subscription), subscription_name));
 }
 #endif